+++ src/App.js
@@ -6,16 +6,16 @@ import { onAuthStateChanged } from 'firebase/auth';
 import { doc, getDoc } from 'firebase/firestore';
 import styled, { createGlobalStyle } from 'styled-components';
 import Cooking from './pages/Cooking';
-import ModifyRecipe from './pages/ModifyRecipe/ModifyRecipe';
-import ReadRecipe from './pages/ReadRecipe/ReadRecipe';
-import SearchRecipe from './pages/SearchRecipe/SearchRecipe';
-import AuthHome from './pages/AuthHome/AuthHome';
-import Register from './pages/Register/Register';
-import UnAuthHome from './pages/UnAuthHome/UnAuthHome';
-import Login from './pages/LogIn/LogIn';
+import ModifyRecipe from './pages/ModifyRecipe';
+import ReadRecipe from './pages/ReadRecipe';
+import SearchRecipe from './pages/SearchRecipe';
+import AuthHome from './pages/AuthHome';
+import Register from './pages/Register';
+import UnAuthHome from './pages/UnAuthHome';
+import Login from './pages/LogIn';
 import { auth, db } from './firestore/index';
 import AuthContext from './components/AuthContext';
-import NoMatch from './pages/NoMatch/NoMatch';
+import NoMatch from './pages/NoMatch';
 import Footer from './components/Footer';
 import ScrollToTop from './components/ScrollToTop';
